requirements.txt
argcomplete>=1.8.1
asn1crypto>=0.24.0
astroid>=2.2.5
awscli>=1.16.192
bcrypt>=3.1.6
beautifulsoup4>=4.8.1
blinker>=1.4
botocore>=1.12.182
Brlapi>=0.6.7
certifi>=2019.6.16
chardet>=3.0.4
chrome-gnome-shell>=0.0.0
colorama>=0.3.9
command-not-found>=0.3
cryptography>=2.6.1
cupshelpers>=1.0
dbus-python>=1.2.12
defer>=1.0.6
distro-info>==0.21ubuntu4
docutils>=0.14
duplicity>=0.8.4
entrypoints>=0.3
et-xmlfile>=1.0.1
fasteners>=0.12.0
fpdf>=1.7.2
future>=0.16.0
httplib2>=0.11.3
idna>=2.8
isort>=4.3.15
jdcal>=1.4.1
jmespath>=0.9.4
keyring>=18.0.1
keyrings.alt>=3.1.1
language-selector>=0.1
launchpadlib>=1.10.7
lazr.restfulclient>=0.14.2
lazr.uri>=1.0.3
lazy-object-proxy>=1.3.1
lockfile>=0.12.2
louis>=3.10.0
macaroonbakery>=1.2.3
MarkupSafe>=1.1.0
mccabe>=0.6.1
monotonic>=1.5
netifaces>=0.10.4
numpy>=1.16.4
oauthlib>=2.1.0
olefile>=0.46
openpyxl>=2.6.2
pandas>=0.24.2
paramiko>=2.6.0
pexpect>=4.6.0
Pillow>=6.1.0
progressbar>=2.5
protobuf>=3.6.1
pyasn1>=0.4.5
pycairo>=1.16.2
pycrypto>=2.6.1
pycups>=1.9.73
PyGObject>=3.34.0
PyJWT>=1.7.0
pylint>=2.3.1
pymacaroons>=0.13.0
PyNaCl>=1.3.0
pyOpenSSL>=19.0.0
pyRFC3339>=1.1
python-apt>=1.9.0+ubuntu1
python-dateutil>=2.7.3
python-debian>=0.1.36
python-gnupg>=0.4.3
pytz>=2019.2
pyxdg>=0.25
PyYAML>=5.1.2
reportlab>=3.5.23
requests>=2.21.0
requests-unixsocket>=0.1.5
rsa>=3.4.2
s3transfer>=0.2.0
simplejson>=3.16.0
six>=1.12.0
soupsieve>=1.9.4
system-service>=0.3
systemd-python>=234
typed-ast>=1.3.1
ubuntu-advantage-tools>=19.5
ubuntu-drivers-common>=0.0.0
Ubuntu-Make>=19.6
ufw>=0.36
unattended-upgrades>=0.1
urllib3>=1.24.1
usb-creator>=0.3.7
wadllib>=1.3.3
wrapt>=1.11.1
xkit>=0.0.0
xlrd>=1.2.0
xlsx2csv>=0.7.6
